Move validators to respective models
