+++ src/main/scala/service/directives/Validator.scala
@@ -8,24 +8,3 @@ trait Validator[T] {
 
 }
 
-object CreateInquestValidator extends Validator[CreateInquest] {
-
-  def validate(createInquest: CreateInquest): Option[ApiError] = {
-    if (createInquest.title.isEmpty)
-      Some(ApiError.invalidInquestTitle(""))
-    else
-      None
-  }
-
-}
-
-object UpdateInquestValidator extends Validator[UpdateInquest] {
-
-  def validate(updateInquest: UpdateInquest): Option[ApiError] = {
-    if (updateInquest.title.exists(_.isEmpty))
-      Some(ApiError.invalidInquestTitle(""))
-    else
-      None
-  }
-
-}
