+++ src/main/scala/service/models/UpdateInquest.scala
@@ -1,6 +1,7 @@
 package service.models
 
 import db.models.Inquest
+import service.directives.Validator
 
 case class UpdateInquest(title: Option[String], description: Option[String]) {
 
@@ -9,5 +10,16 @@ case class UpdateInquest(title: Option[String], description: Option[String]) {
       title = title.getOrElse(inquest.title),
       description = description.getOrElse(inquest.description)
     )
+
+}
+
+object UpdateInquestValidator extends Validator[UpdateInquest] {
+
+  def validate(updateInquest: UpdateInquest): Option[ApiError] = {
+    if (updateInquest.title.exists(_.isEmpty))
+      Some(ApiError.invalidInquestTitle(""))
+    else
+      None
+  }
 
 }
